refactor(treasury): factor bootstrapRunLoC out of test-runLoC
